% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopyInterception.R
\name{canopyInterception}
\alias{canopyInterception}
\title{Snow canopy interception model (Hedstrom and Pomeroy, 1998)}
\usage{
canopyInterception(preType, Tair, LAI, P, Smean, L0, c, MaxCanopyStorage)
}
\arguments{
\item{preType}{() precipitation type TRUE = snow, FALSE = rain}

\item{Tair}{(deg C) air temperature}

\item{LAI}{()      Leaf area index}

\item{P}{(mm)    precipitation}

\item{Smean}{(kg/m^2)  snow loading coefficient}

\item{L0}{(mm SWE) initial snow load}

\item{c}{() dimensionless unloading coefficient = 0.678}

\item{MaxCanopyStorage}{(mm/day) changes according to LAI values}
}
\value{
effective rain/snow (mm)
}
\description{
THis function compute evapotranspiration based on radiation, conductance etc
}
\author{
Bar Avni
}

