\name{EcohydrotoolsBAR3-package}
\alias{EcohydrotoolsBAR3-package}
\alias{EcohydrotoolsBAR3}
\docType{package}
\title{
\packageTitle{EcohydrotoolsBAR3}
}
\description{
\packageDescription{EcohydrotoolsBAR3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EcohydrotoolsBAR3}
\packageIndices{EcohydrotoolsBAR3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EcohydrotoolsBAR3}

Maintainer: \packageMaintainer{EcohydrotoolsBAR3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
