% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evapo_HS.R
\name{Evapo_HS}
\alias{Evapo_HS}
\title{Hargreaves and Samani (1985) simple evapotranspiration model}
\usage{
Evapo_HS(date, lat, LAI_living, LAI, Tave, Tmax, Tmin)
}
\arguments{
\item{date}{()}

\item{lat}{(deg) Latitude in degrees}

\item{LAI_living}{() LAI for healthy forest}

\item{LAI}{() current stage LAI}

\item{Tave}{(deg C) average temperature per day}

\item{Tmax}{(deg C) maximum temperature per day}

\item{Tmin}{(deg C) minimum temperature per day}
}
\value{
evapotranspiration (mm/day)
}
\description{
This function calculates the daily evapotranspiration estimation using the Hargreaves and Samani 1985 method
}
\author{
Bar Avni
}

