% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExRadiation.R
\name{ExRadiation}
\alias{ExRadiation}
\title{extraterrestrial solar radiation model}
\usage{
ExRadiation(date, lat, LAI_living, LAI)
}
\arguments{
\item{date}{()}

\item{lat}{(deg) Latitude in degrees}

\item{LAI_living}{() LAI for healthy forest}

\item{LAI}{() current stage LAI}
}
\value{
Extraterrestrial radiation (MJ m^-2)day^-1
}
\description{
This function calculates the daily external radiation following the formulation presented in Allen et al, 1988
Date is converted to J is the number of the day in the year between 1 (1 January) and 365 or 366 (31 December).
}
\author{
Bar Avni
}

