% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoffCalc.R, R/runoffCalc2.R
\name{runoffCalc}
\alias{runoffCalc}
\title{Runoff model}
\usage{
runoffCalc(P, stage, MaxStorage = 25, ET, year)

runoffCalc(P, stage, MaxStorage = 25, ET, year)
}
\arguments{
\item{P}{(mm SWE) monthly snow precipitation}

\item{MaxStorage}{(mm) maximum capacity of ground storage}

\item{ET}{(mm/day) evapotranspiration}

\item{year}{}

\item{Q}{(mm) effective rain/snow (mm)}

\item{InfilCap}{(mm/day) constant infiltration capacity}

\item{month}{}

\item{I}{(mm) monthly canopy interception}

\item{InfilCap}{(mm/month) constant infiltration capacity}

\item{MaxStorage}{(mm) maximum capacity of ground storage}

\item{ET}{(mm/month) evapotranspiration}

\item{year}{}
}
\value{
monthly runoff depth (mm)

list of yearly runoff depth (mm) for three lodgpol pine mortality stages - living, red, grey (source: Pugh and Gordon, 2013)
}
\description{
THis function compute runoff from evapotranspiration, infiltration and snow precipitation

THis function compute runoff from evapotranspiration, infiltration and snow precipitation
}
\author{
Bar Avni

Bar Avni
}

