% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Water_yield_cold_env.R
\name{Water_yield_cold_env}
\alias{Water_yield_cold_env}
\title{Runoff model}
\usage{
Water_yield_cold_env(LAI_living, LAI, MS, preType, Tair, P, Smean, L0, c, date,
  lat, Tave, Tmax, Tmin, InfilCap, MaxStorage, month, year)
}
\arguments{
\item{LAI_living}{() LAI for healthy forest}

\item{LAI}{()      Leaf area index}

\item{MS}{(mm/day) maximum rain interception in a living forest}

\item{preType}{() precipitation type TRUE = snow, FALSE = rain}

\item{Tair}{(deg C) air temperature}

\item{P}{(mm)    daily precipitation}

\item{Smean}{(kg/m^2)  snow loading coefficient}

\item{L0}{(mm SWE) initial snow load}

\item{c}{() dimensionless unloading coefficient = 0.678}

\item{date}{()}

\item{lat}{(deg) Latitude in degrees}

\item{Tave}{(deg C) average temperature per day}

\item{Tmax}{(deg C) maximum temperature per day}

\item{Tmin}{(deg C) minimum temperature per day}

\item{InfilCap}{(mm/day) constant infiltration capacity}

\item{MaxStorage}{(mm) maximum capacity of ground storage}

\item{month}{}

\item{year}{}
}
\value{
monthly runoff depth (mm)
}
\description{
THis function compute runoff from evapotranspiration, infiltration and snow precipitation
}
\author{
Bar Avni
}

