% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoffCalc.R
\name{runoffCalc}
\alias{runoffCalc}
\title{Runoff model}
\usage{
runoffCalc(Q, InfilCap, MaxStorage, ET, month, year)
}
\arguments{
\item{Q}{(mm) effective rain/snow (mm)}

\item{InfilCap}{(mm/day) constant infiltration capacity}

\item{MaxStorage}{(mm) maximum capacity of ground storage}

\item{ET}{(mm/day) evapotranspiration}

\item{month}{}

\item{year}{}
}
\value{
monthly runoff depth (mm)
}
\description{
THis function compute runoff from evapotranspiration, infiltration and snow precipitation
}
\author{
Bar Avni
}

